add test execution to command
