+++ internal/test/resolvers.go
@@ -1,3 +1,21 @@
+/*
+ * Flow CLI
+ *
+ * Copyright 2022 Dapper Labs, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package test
 
 import (
@@ -12,7 +30,7 @@ import (
 func importResolver(
 	scriptPath string,
 	readerWriter flowkit.ReaderWriter,
-	contracts config.Contracts,
+	contracts *config.Contracts,
 ) cdcTests.ImportResolver {
 	return func(location common.Location) (string, error) {
 		stringLocation, isFileImport := location.(common.StringLocation)
@@ -37,11 +55,11 @@ func importResolver(
 }
 
 func resolveContract(
-	contracts config.Contracts,
+	contracts *config.Contracts,
 	stringLocation common.StringLocation,
 ) (config.Contract, error) {
 	relativePath := stringLocation.String()
-	for _, contract := range contracts {
+	for _, contract := range *contracts {
 		if contract.Location == relativePath {
 			return contract, nil
 		}
